requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
rembg[cpu]>=2.0.50
mediapipe>=0.10.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
